webserver/requirements.txt
python-dotenv
requests
beautifulsoup4
openai
pillow
py3-tts
bottle